+++ app/src/main/java/com/sopt/clody/presentation/ui/auth/login/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.sopt.clody.presentation.ui.auth.signup
+package com.sopt.clody.presentation.ui.auth.login
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -25,42 +25,37 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.sopt.clody.R
 import com.sopt.clody.presentation.ui.auth.component.button.KaKaoButton
-import com.sopt.clody.presentation.ui.auth.navigation.AuthNavigator
+import com.sopt.clody.presentation.ui.auth.signup.SignUpViewModel
 import com.sopt.clody.presentation.ui.component.LoadingScreen
 import com.sopt.clody.presentation.utils.base.UiState
 import com.sopt.clody.presentation.utils.extension.heightForScreenPercentage
 import com.sopt.clody.ui.theme.ClodyTheme
 
 @Composable
-fun SignUpRoute(
-    authNavigator: AuthNavigator,
+fun LoginRoute(
+    navigateToTerms: () -> Unit,
+    navigateToHome: () -> Unit,
 ) {
     val viewModel: SignUpViewModel = hiltViewModel()
     val signInState by viewModel.signInState.collectAsState()
     val context = LocalContext.current
 
-    LaunchedEffect(signInState) {
+    LaunchedEffect(signInState.uiState) {
         when (signInState.uiState) {
-            is UiState.Success -> {
-                authNavigator.navigateHome()
-            }
-
-            is UiState.Failure -> {
-                authNavigator.navigateTermsOfService()
-            }
-
-            else -> {}
+            is UiState.Success -> navigateToHome()
+            is UiState.Failure -> navigateToTerms()
+            else -> Unit
         }
     }
 
-    SignUpScreen(
+    LoginScreen(
         isLoading = signInState.uiState is UiState.Loading,
         onSignInClick = { viewModel.signInWithKakao(context) },
     )
 }
 
 @Composable
-fun SignUpScreen(
+fun LoginScreen(
     isLoading: Boolean,
     onSignInClick: () -> Unit,
 ) {
@@ -85,34 +80,33 @@ fun SignUpScreen(
                     .padding(bottom = 40.dp),
             )
         },
-        content = { innerPadding ->
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(color = backgroundColor)
-                    .padding(innerPadding),
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                Spacer(modifier = Modifier.heightForScreenPercentage(0.38f))
-                Image(
-                    painter = painterResource(id = R.drawable.ic_signup_logo),
-                    contentDescription = null,
-                    contentScale = ContentScale.Crop,
-                )
-                Spacer(modifier = Modifier.heightForScreenPercentage(0.02f))
-                Image(
-                    painter = painterResource(id = R.drawable.ic__signup_title),
-                    contentDescription = null,
-                )
-                Spacer(modifier = Modifier.heightForScreenPercentage(0.01f))
-                Image(
-                    painter = painterResource(id = R.drawable.ic_signup_logotitle),
-                    contentDescription = null,
-                    contentScale = ContentScale.Crop,
-                )
-            }
-        },
-    )
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(color = backgroundColor)
+                .padding(innerPadding),
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
+            Spacer(modifier = Modifier.heightForScreenPercentage(0.38f))
+            Image(
+                painter = painterResource(id = R.drawable.ic_signup_logo),
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+            )
+            Spacer(modifier = Modifier.heightForScreenPercentage(0.02f))
+            Image(
+                painter = painterResource(id = R.drawable.ic__signup_title),
+                contentDescription = null,
+            )
+            Spacer(modifier = Modifier.heightForScreenPercentage(0.01f))
+            Image(
+                painter = painterResource(id = R.drawable.ic_signup_logotitle),
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+            )
+        }
+    }
 
     if (isLoading) {
         LoadingScreen()
@@ -121,8 +115,8 @@ fun SignUpScreen(
 
 @Preview(showBackground = true)
 @Composable
-fun RegisterScreenPreview() {
-    SignUpScreen(
+fun LoginScreenPreview() {
+    LoginScreen(
         isLoading = false,
         onSignInClick = {},
     )
